improve error handling
this improves error handling by not using `.parse` everwhere which
results in a lot of throws without a reference to which input it came
from.
now `.safeParse` is used and each input has an individual error check.
additionally the 2 inputs which get JSON.parse'ed are now also checked
for malformed json beforehand.
an error message has also been added to `semverSchema`